+++ ios/IJKMediaPlayer/IJKMediaPlayer/IJKFFMrl.m
@@ -79,7 +79,7 @@
 
     if (out_size > INT_MAX || !(utf8Plain = av_mallocz(out_size)))
         return NULL;
-    if ((ret = av_base64_decode(utf8Plain, utf8Cipher, out_size)) < 0) {
+    if ((ret = av_base64_decode(utf8Plain, utf8Cipher, (int)out_size)) < 0) {
         av_free(utf8Plain);
         NSLog(@"Invalid base64 in MRL\n");
         return NULL;
